AG-1876: move escapeRegexChars to shared lib so it can be reused by agora
