--- apps/model-ad/api/src/utils/regex.ts
@@ -1,4 +1,0 @@
-export function escapeRegexChars(input: string): string {
-  // Escape all regex special characters to prevent injection
-  return input.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, '\\$&');
-}
